+++ src/core/ParserFunc.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import type { TParamMapIn, TTextMapIn, TValMapIn } from '../AhkSymbol/CAhkFunc';
 import { CAhkFunc } from '../AhkSymbol/CAhkFunc';
 import type { TTokenStream } from '../globalEnum';
+import { EFnMode } from '../tools/DeepAnalysis/FnVar/EFnMode';
 import { getFnVarDef } from '../tools/DeepAnalysis/FnVar/getFnVarDef';
 import { getParamDef } from '../tools/DeepAnalysis/getParamDef';
 import { getUnknownTextMap } from '../tools/DeepAnalysis/getUnknownTextMap';
@@ -81,7 +82,7 @@ export function getFunc(FuncInput: TFuncInput): CAhkFunc | null {
     // if is static mode
     const allowList: readonly boolean[] = getAllowsListOfFunc(DocStrMap, startLine, endLine);
 
-    const valMap: TValMapIn = getFnVarDef(allowList, AhkTokenList, paramMap, GValMap);
+    const valMap: TValMapIn = getFnVarDef(allowList, AhkTokenList, paramMap, GValMap, EFnMode.normal);
     const textMap: TTextMapIn = getUnknownTextMap(allowList, AhkTokenList, paramMap, valMap, GValMap, name); // eval!!
 
     const selectionRangeText: string = document.getText(selectionRange);