+++ src/core/ParserTools/ahkGlobalVarTool.ts
@@ -2,7 +2,7 @@ import { pm } from '../ProjectManager';
 
 type TData = {
     fsPathList: string[];
-    rawName: string;
+    keyRawName: string;
     // and any you want to has Data
 };
 
@@ -11,8 +11,8 @@ export type TGlobalVarUpNameMap = ReadonlyMap<string, TData>;
 export function getGlobalUpNameMap(): TGlobalVarUpNameMap {
     const map = new Map<string, TData>();
 
-    for (const { GValMap, uri } of pm.getDocMapValue()) {
-        for (const [k, { rawName }] of GValMap) {
+    for (const { ModuleVar, uri } of pm.getDocMapValue()) {
+        for (const [k, { keyRawName }] of ModuleVar.ModuleValMap) {
             const oldData: TData | undefined = map.get(k);
             const fsPathList: string[] = oldData !== undefined
                 ? [...oldData.fsPathList, uri.toString()]
@@ -20,7 +20,7 @@ export function getGlobalUpNameMap(): TGlobalVarUpNameMap {
 
             map.set(k, {
                 fsPathList,
-                rawName,
+                keyRawName,
             });
         }
         //