+++ src/provider/Hover/tools/hoverGlobalVar.ts
@@ -1,14 +1,12 @@
 /* eslint-disable max-statements */
 import path from 'node:path';
 import * as vscode from 'vscode';
-import type { TGlobalVal } from '../../../core/ParserTools/ahkGlobalDef';
 import { pm } from '../../../core/ProjectManager';
 
 export function hoverGlobalVar(wordUp: string): vscode.MarkdownString | null {
     const msg: string[] = [];
-    for (const { GValMap, uri } of pm.getDocMapValue()) {
-        const gVal: Readonly<TGlobalVal> | undefined = GValMap.get(wordUp);
-        if (gVal !== undefined) {
+    for (const { ModuleVar, uri } of pm.getDocMapValue()) {
+        if (ModuleVar.ModuleValMap.has(wordUp)) {
             const fsPath: string = uri.toString();
             msg.push(`- [${path.basename(fsPath)}](${fsPath})`);
         }