+++ src/provider/SemanticTokensProvider/SemanticTokensProvider.ts
@@ -4,7 +4,6 @@ import { showTimeSpend } from '../../configUI';
 import type { TAhkFileData } from '../../core/ProjectManager';
 import { pm } from '../../core/ProjectManager';
 import { DAList2SemanticHighlight } from './DAList2SemanticHighlight';
-import { GlobalHighlight } from './GlobalHighlight';
 import { ModuleVarSemantic } from './ModuleVarSemantic';
 import { pushToken, TokenModifiers, TokenTypes } from './tools';
 // https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide#standard-token-types-and-modifiers
@@ -22,17 +21,12 @@ function SemanticTokensCore(document: vscode.TextDocument): vscode.SemanticToken
     const cache: vscode.SemanticTokens | undefined = wm.get(AhkFileData);
     if (cache !== undefined) return cache;
 
-    const {
-        GValMap,
-        AST,
-        ModuleVar,
-    } = AhkFileData;
+    const { AST, ModuleVar } = AhkFileData;
 
     const tokensBuilder: vscode.SemanticTokensBuilder = new vscode.SemanticTokensBuilder(legend);
 
     pushToken([
         ...DAList2SemanticHighlight(AST),
-        ...GlobalHighlight(GValMap),
         ...ModuleVarSemantic(ModuleVar),
     ], tokensBuilder);