+++ src/tools/DeepAnalysis/FnVar/TFnVarDef.ts
@@ -1,5 +1,6 @@
 import type { TParamMapIn, TValMapIn } from '../../../AhkSymbol/CAhkFunc';
 import type { TGValMap } from '../../../core/ParserTools/ahkGlobalDef';
+import type { EFnMode } from './EFnMode';
 
 export type TGetFnDefNeed = {
     lStr: string;
@@ -9,4 +10,5 @@ export type TGetFnDefNeed = {
     GValMap: TGValMap;
     lStrTrimLen: number;
     lineComment: string;
+    fnMode: EFnMode;
 };