+++ src/tools/DeepAnalysis/FnVar/def/getValMeta.ts
@@ -1,13 +1,26 @@
 import * as vscode from 'vscode';
 import type { TValMapIn, TValMetaIn } from '../../../../AhkSymbol/CAhkFunc';
+import type { EFnMode } from '../EFnMode';
 import { wrapFnValDef } from './wrapFnValDef';
 
+type TValMeta = {
+    line: number;
+    character: number;
+    RawName: string;
+    valMap: TValMapIn;
+    lineComment: string;
+    fnMode: EFnMode;
+};
+
 export function getValMeta(
-    line: number,
-    character: number,
-    RawName: string,
-    valMap: TValMapIn,
-    lineComment: string,
+    {
+        line,
+        character,
+        RawName,
+        valMap,
+        lineComment,
+        fnMode,
+    }: TValMeta,
 ): TValMetaIn {
     const defRange: vscode.Range = new vscode.Range(
         new vscode.Position(line, character),
@@ -19,6 +32,7 @@ export function getValMeta(
         valMap,
         defRange,
         lineComment,
+        fnMode,
     });
     return value;
 }