+++ src/tools/DeepAnalysis/FnVar/getFnVarDef.ts
@@ -1,19 +1,71 @@
+/* eslint-disable max-statements */
+/* eslint-disable max-lines-per-function */
 import type { TParamMapIn, TValMapIn, TValMetaIn } from '../../../AhkSymbol/CAhkFunc';
 import type { TGValMap } from '../../../core/ParserTools/ahkGlobalDef';
+import type { TVarData } from '../../../core/ParserTools/varMixedAnnouncement';
+import { varMixedAnnouncement } from '../../../core/ParserTools/varMixedAnnouncement';
 import type { TTokenStream } from '../../../globalEnum';
+import { replacerSpace } from '../../str/removeSpecialChar';
 import { forLoop } from './def/forLoop';
+import { getValMeta } from './def/getValMeta';
 import { OutputVarCommandBase } from './def/OutputVarCommandBase';
 import { OutputVarCommandPlus } from './def/OutputVarCommandPlus';
 import { varSetCapacityFunc } from './def/varSetCapacityFunc';
 import { walrusOperator } from './def/walrusOperator';
+import { EFnMode } from './EFnMode';
 import type { TGetFnDefNeed } from './TFnVarDef';
+
+type TParam = {
+    varDataList: TVarData[];
+    line: number;
+    valMap: TValMapIn;
+    lineComment: string;
+    fistWordVarMix: 'LOCAL' | 'STATIC';
+};
+function setVarMix({
+    varDataList,
+    line,
+    valMap,
+    lineComment,
+    fistWordVarMix,
+}: TParam): void {
+    for (const { ch, rawName } of varDataList) {
+        const value: TValMetaIn = getValMeta(
+            {
+                line,
+                character: ch,
+                RawName: rawName,
+                valMap,
+                lineComment,
+                fnMode: fistWordVarMix === 'LOCAL'
+                    ? EFnMode.forceLocal
+                    : EFnMode.static,
+            },
+        );
+        valMap.set(rawName.toUpperCase(), value);
+    }
+}
+
+function setFnMode(fistWordUp: string, lStrTrimLen: number, fnMode: EFnMode): EFnMode {
+    if (fistWordUp === 'STATIC' && lStrTrimLen > 'STATIC'.length) {
+        return EFnMode.static;
+    }
+    if (fistWordUp === 'LOCAL' && lStrTrimLen > 'LOCAL'.length) {
+        return EFnMode.forceLocal;
+    }
+    return fnMode;
+}
 
 export function getFnVarDef(
     allowList: readonly boolean[],
     DocStrMap: TTokenStream,
     paramMap: TParamMapIn,
     GValMap: TGValMap,
+    fnModeDefault: EFnMode,
 ): TValMapIn {
+    let fnMode: EFnMode = fnModeDefault;
+    let fistWordVarMix: '' | 'GLOBAL' | 'LOCAL' | 'STATIC' = '';
+
     const valMap: TValMapIn = new Map<string, TValMetaIn>();
     for (
         const {
@@ -21,6 +73,7 @@ export function getFnVarDef(
             line,
             lineComment,
             fistWordUp,
+            cll,
         } of DocStrMap
     ) {
         if (!allowList[line]) continue; // in arg Range
@@ -28,6 +81,39 @@ export function getFnVarDef(
 
         const lStrTrimLen: number = lStr.trim().length;
         if (lStrTrimLen < 2) continue; // a=b need length >=3
+
+        if (fnMode !== EFnMode.normal) {
+            fnMode = setFnMode(fistWordUp, lStrTrimLen, fnMode);
+        }
+
+        if (
+            lStrTrimLen > 'LOCAL'.length
+            && (fistWordUp === 'STATIC' || fistWordUp === 'LOCAL' || fistWordUp === 'GLOBAL')
+        ) {
+            fistWordVarMix = fistWordUp;
+        } else if (fistWordVarMix !== '' && cll === 1) {
+            // nothing
+            // fistWordVarMix = the last line fistWordUp
+        } else {
+            fistWordVarMix = '';
+        }
+
+        if (fistWordVarMix === 'GLOBAL') continue;
+
+        if (fistWordVarMix === 'LOCAL' || fistWordVarMix === 'STATIC') {
+            const varDataList = varMixedAnnouncement(lStr.replace(/^\s*\b(?:static|local)\b[,\s]+/ui, replacerSpace));
+            if (varDataList.length > 0) {
+                console.log('🚀 varDataList', varDataList, lStr);
+            }
+            setVarMix({
+                varDataList,
+                line,
+                valMap,
+                lineComment,
+                fistWordVarMix,
+            });
+            continue;
+        }
 
         const need: TGetFnDefNeed = {
             lStr,
@@ -37,6 +123,7 @@ export function getFnVarDef(
             GValMap,
             lStrTrimLen,
             lineComment,
+            fnMode,
         };
         walrusOperator(need); // :=
         varSetCapacityFunc(need); // VarSetCapacity(varName) or NumGet(varName) or NumGet(&varName)